MSVC: fix the /W4 /WX conversions; add a full-warning CI diagnostic build
The Windows legs failed on error C2220 in src/core/Arcade.cpp (a double->float
conversion promoted to an error). Fixes the MSVC-level-4 class:
- kMonoCharWidthFactor is a float now (it multiplies float font sizes)
- the rhythm falling-note Y keeps the approach time in float
- stats.meter conversions are explicit
- arcade_bench's FNV-1a loop no longer converts char -> unsigned char implicitly

CMake learned KIEEKEY_WERROR (default ON, so CI keeps /WX) and the Build step
now, when the strict build fails, rebuilds once with KIEEKEY_WERROR=OFF and
re-emits every unique diagnostic as an annotation: /WX stops each translation
unit at its first warning, so a single warning per file would otherwise hide the
rest of the fix list.

+++ src/core/Arcade.cpp
@@ -41,7 +41,7 @@ namespace ok::arcade {
 namespace {
 
 constexpr double kMaxFrameStepSec = 0.05;   // 20 FPS floor for physics stepping
-constexpr double kMonoCharWidthFactor = 0.62;  // monospace advance / font size
+constexpr float kMonoCharWidthFactor = 0.62f;  // monospace advance / font size
 
 double clampDt(double dt) noexcept {
     if (!(dt > 0.0)) {   // also filters NaN
@@ -2032,7 +2032,9 @@ void RhythmTypingGame::buildFrame(Frame& frame) const {
         if (dtToHit > m_approachSec || dtToHit < -kLateWindowSec) {
             continue;
         }
-        const float y = kHitLineY - static_cast<float>(dtToHit) * (kPixelsPerSecond / m_approachSec);
+        const float y = kHitLineY -
+                       static_cast<float>(dtToHit) *
+                           (kPixelsPerSecond / static_cast<float>(m_approachSec));
         if (y < 50.0f || y > kHitLineY + 30.0f) {
             continue;
         }
@@ -2300,7 +2302,7 @@ void NoMistakeGame::buildFrame(Frame& frame) const {
                                : static_cast<double>(m_currentIndex) /
                                      static_cast<double>(m_textStream.size());
     frame.stats.hasMeter = true;
-    frame.stats.meter = m_score;
+    frame.stats.meter = static_cast<double>(m_score);
     frame.stats.meterMax = static_cast<double>(m_scorePenalty);
     frame.stats.paused = m_paused;
     frame.stats.gameOver = m_gameOver;
